refactor: HttpRequest의 uri, method를 RequestLine으로 분리
